+++ react/ExtensionContainer.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { ExtensionContainer } from 'vtex.render-runtime'
 
 export default ExtensionContainer