+++ react/LayoutContainer.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { LayoutContainer } from 'vtex.render-runtime'
 
 export default LayoutContainer