+++ react/LegacyExtensionContainer.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { LegacyExtensionContainer } from 'vtex.render-runtime'
 
 export default LegacyExtensionContainer