+++ react/error.tsx
@@ -1,6 +1,7 @@
 /* global module */
 import React, { Component, Fragment } from 'react'
 import ReactJson from 'react-json-view'
+import { captureException } from '@sentry/react'
 
 require('myvtex-sse')
 
@@ -10,15 +11,32 @@ import ErrorImg from './images/error-img.png'
 
 import style from './error.css'
 import { renderReadyPromise } from '.'
+import { isAdmin } from './utils/isAdmin'
 
 class ErrorPage extends Component {
   public state = { enabled: false }
-  private splunk = 0
 
   public componentDidMount() {
     window.setTimeout(() => {
       this.setState({ enabled: true })
     }, 5000)
+
+    if (!isAdmin()) {
+      return
+    }
+
+    const error = window?.__ERROR__
+    const requestId = window?.__REQUEST_ID__
+    const defaultError =
+      'Render Runtime renderered an error page and there is no error or request id available'
+
+    if (error) {
+      captureException(error)
+    } else if (requestId) {
+      captureException(requestId)
+    } else {
+      captureException(defaultError)
+    }
   }
 
   public render() {