+++ react/index.tsx
@@ -1,4 +1,5 @@
 /* global module */
+import './o11y/instrument'
 import 'core-js/es6/symbol'
 import 'core-js/fn/symbol/iterator'
 import { canUseDOM } from 'exenv'
@@ -12,7 +13,10 @@ import { loadRuntimeJSONs } from './start/runtime'
 import { hydrateUncriticalStyles } from './start/styles'
 
 function performanceMark(...args: Parameters<typeof window.performance.mark>) {
-  if (typeof window === 'undefined' || !(typeof window?.performance?.mark === 'function')) {
+  if (
+    typeof window === 'undefined' ||
+    !(typeof window?.performance?.mark === 'function')
+  ) {
     return
   }
   window.performance.mark(...args)
@@ -46,7 +50,6 @@ export const renderReadyPromise: Promise<any> = canUseDOM
       contentLoadedPromise.then(() => {
         performanceMark('content-loaded-promise-resolved')
       })
-
 
       const scriptsLoadedPromise = new Promise((resolve) => {
         if (typeof window.__ASYNC_SCRIPTS_READY__ === 'undefined') {